timeline: Allow disabling seek when left-clicking
Currently the user can select, edit and seek with the left mouse button.
The behavior of the right mouse button has not been not defined yet.
It does some stuff similarly with the LMB and other times it does nothing.

With this change, the RMB can be used only for seeking. Now it's also possible
to scrub with the RMB directly on the timeline. Previously this was possible
only on the ruler (with LMB or RMB).

Now it's possible to disable seeking when the LMB is clicked. There is a new
"Left click also seeks" preference which is enabled by default because
this is the old behavior.

Fixes https://phabricator.freedesktop.org/T3158