widgets: Use Grid for GstElementSettingsWidget
Also change the widgets for editing properties so they don't expand.
This was a bit difficult with NumericWidget because it uses a SpinButton
and when the max value of the spin button's adjustment is very large it
takes up a lot of space. Fix this by setting a cap on the limits.